requirements.txt
mysql-connector-python==8.0.33
pymysql==1.0.3
python-dotenv==1.0.0
